+++ matchbox-server/src/main/java/ch/ahdis/matchbox/packages/ImplementationGuideProviderR5.java
@@ -554,4 +554,9 @@ public class ImplementationGuideProviderR5 extends ImplementationGuideResourcePr
 				.setVersion(packageVersion)
 		);
 	}
+
+	@Override
+	public long count() {
+		return this.myPackageVersionDao.count();
+	}
 }